front: separate useLazyProjectTrains and useLazyLoadTrains hooks
